+++ src/sst/core/componentInfo.cc
@@ -13,8 +13,8 @@
 
 #include "sst/core/componentInfo.h"
 
-#include "sst/core/configGraph.h"
 #include "sst/core/linkMap.h"
+#include "sst/core/model/configGraph.h"
 #include "sst/core/serialization/serialize.h"
 #include "sst/core/serialization/serializer.h"
