+++ src/sst/core/configGraphOutput.h
@@ -12,7 +12,7 @@
 #ifndef SST_CORE_CONFIGGRAPH_OUTPUT_H
 #define SST_CORE_CONFIGGRAPH_OUTPUT_H
 
-#include "sst/core/configGraph.h"
+#include "sst/core/model/configGraph.h"
 #include "sst/core/params.h"
 #include "sst/core/util/filesystem.h"
