+++ src/sst/core/model/Makefile.inc
@@ -8,10 +8,31 @@
 sst_core_sources += \
 	model/sstmodel.h \
 	model/sstmodel.cc \
+	model/configComponent.h \
+	model/configComponent.cc \
+	model/configGraph.cc \
+	model/configLink.h \
+	model/configLink.cc \
+	model/configStatistic.h \
+	model/configStatistic.cc \
 	model/element_python.h \
 	model/element_python.cc \
 	model/restart/sstcptmodel.h \
-	model/restart/sstcptmodel.cc
+	model/restart/sstcptmodel.cc \
+	model/cfgoutput/pythonConfigOutput.h \
+	model/cfgoutput/pythonConfigOutput.cc \
+	model/cfgoutput/dotConfigOutput.h \
+	model/cfgoutput/dotConfigOutput.cc \
+	model/cfgoutput/xmlConfigOutput.h \
+	model/cfgoutput/xmlConfigOutput.cc \
+	model/cfgoutput/jsonConfigOutput.h \
+	model/cfgoutput/jsonConfigOutput.cc
+
+
+nobase_dist_sst_HEADERS += \
+	model/configGraph.h \
+	model/element_python.h
+
 
 sst_core_python_headers = \
   model/python/pymacros.h
