+++ src/sst/core/model/cfgoutput/pythonConfigOutput.cc
@@ -12,7 +12,7 @@
 
 #include "sst_config.h"
 
-#include "pythonConfigOutput.h"
+#include "sst/core/model/cfgoutput/pythonConfigOutput.h"
 
 #include "sst/core/config.h"
 #include "sst/core/simulation_impl.h"
