+++ src/sst/core/model/python/pymodel_comp.cc
@@ -17,7 +17,7 @@
 
 #include "sst/core/component.h"
 #include "sst/core/componentInfo.h"
-#include "sst/core/configGraph.h"
+#include "sst/core/model/configGraph.h"
 #include "sst/core/model/python/pymacros.h"
 #include "sst/core/model/python/pymodel.h"
 #include "sst/core/model/python/pymodel_link.h"
