+++ src/sst/core/portModule.cc
@@ -14,7 +14,7 @@
 #include "sst/core/portModule.h"
 
 #include "sst/core/baseComponent.h"
-#include "sst/core/configGraph.h"
+#include "sst/core/model/configGraph.h"
 #include "sst/core/simulation_impl.h"
 #include "sst/core/stringize.h"
 #include "sst/core/warnmacros.h"
