+++ src/sst/core/simulation.cc
@@ -17,7 +17,6 @@
 #include "sst/core/checkpointAction.h"
 #include "sst/core/clock.h"
 #include "sst/core/config.h"
-#include "sst/core/configGraph.h"
 #include "sst/core/exit.h"
 #include "sst/core/factory.h"
 #include "sst/core/heartbeat.h"
@@ -25,6 +24,7 @@
 #include "sst/core/interactiveConsole.h"
 #include "sst/core/linkMap.h"
 #include "sst/core/linkPair.h"
+#include "sst/core/model/configGraph.h"
 #include "sst/core/output.h"
 #include "sst/core/profile/clockHandlerProfileTool.h"
 #include "sst/core/profile/eventHandlerProfileTool.h"
