+++ src/sst/core/statapi/statgroup.cc
@@ -14,7 +14,7 @@
 #include "sst/core/statapi/statgroup.h"
 
 #include "sst/core/baseComponent.h"
-#include "sst/core/configGraph.h"
+#include "sst/core/model/configGraph.h"
 #include "sst/core/output.h"
 #include "sst/core/simulation_impl.h"
 #include "sst/core/statapi/statbase.h"
